Update:

This PR adds AVX 512-bit support. The performance compared to AVX2 is worse. Either I am not utilising correctly the 512-bit instructions set, or it simply does not provide any benefit for this type of computation. I'll leave this draft PR if other people are interested in giving it a try, but for now I am not going to merge it.

---
## OUTDATED BELOW

WIP in progress

This is not tested because I don't have a AVX 512-bit CPU, so very likely that the code will fail.
Still, would appreciate if someone gives it a try and report issues.

@ArtyomZemlyak
Are you interested in giving it a try? I noticed you have a CPU with AVX 512-bit support