@ArtyomZemlyak 2022-10-27
Seems its much slower for all models