@ArtyomZemlyak 2022-10-27
3 runs of tiny -t 8 on AVX512 and master (AVX2) branches
![image](https://user-images.githubusercontent.com/42967602/198222479-13e740cc-09bc-4686-86f3-a6960e65341f.png)