@ArtyomZemlyak 2022-10-27
Cpu info (all):

fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush mmx fxsr sse sse2 ss ht syscall nx pdpe1gb rdtscp lm constant_tsc arch_perfmon rep_good nopl xtopology tsc_reliable nonstop_tsc cpuid pni pclmulqdq vmx ssse3 fma cx16 pdcm pcid sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave `avx` f16c rdrand hypervisor lahf_lm abm 3dnowprefetch invpcid_single ssbd ibrs ibpb stibp ibrs_enhanced tpr_shadow vnmi ept vpid ept_ad fsgsbase tsc_adjust bmi1 `avx2` smep bmi2 erms invpcid `avx512f` `avx512dq` rdseed adx smap `avx512ifma` clflushopt clwb `avx512cd` sha_ni `avx512bw` `avx512vl` xsaveopt xsavec xgetbv1 xsaves `avx512vbmi` umip `avx512_vbmi2` gfni vaes vpclmulqdq `avx512_vnni` `avx512_bitalg` `avx512_vpopcntdq` rdpid movdiri movdir64b fsrm `avx512_vp2intersect` flush_l1d arch_capabilities