@jaybinks 2022-11-06
have re-tested, and it seems to be no better (possibly worse)
./bench -m ./models/ggml-small.en.bin -t 4

before pulling your recent change :
```
whisper_print_timings:     load time =  1096.42 ms
whisper_print_timings:      mel time =     0.00 ms
whisper_print_timings:   sample time =     0.00 ms
whisper_print_timings:   encode time = 11793.09 ms / 982.76 ms per layer
whisper_print_timings:   decode time =     0.00 ms / 0.00 ms per layer
whisper_print_timings:    total time = 12890.37 ms
```

after commit c4350356deafcc748b64f6aece9d4de2cf223de5
```
whisper_print_timings:     load time =   785.93 ms
whisper_print_timings:      mel time =     0.00 ms
whisper_print_timings:   sample time =     0.00 ms
whisper_print_timings:   encode time = 13540.00 ms / 1128.33 ms per layer
whisper_print_timings:   decode time =     0.00 ms / 0.00 ms per layer
whisper_print_timings:    total time = 14325.94 ms
```